refactor: remove inventory-plus references and consolidate inventory views
- Removed old InventoryView.vue
- Renamed InventoryPlusView.vue to InventoryView.vue
- Updated router to remove inventory-plus route
- Updated LeftSideNav to use inventory:index route
- Consolidated all inventory functionality into single view
